+++ packages/necroparser/test/fixtures.test.ts
@@ -1,13 +1,15 @@
+/// <reference types="vite/client" />
+
 import { fileURLToPath, resolve } from 'node:url';
 import { describe, expect, test } from 'vitest';
-import { getInterfaces } from '../src';
+import { getInterfaces } from '../src/index.js';
 
-const fixtures = import.meta.glob('./fixtures/*.ts', {
+const fixturesTest = import.meta.glob('./fixtures/*.ts', {
 	query: '?raw'
 });
 
 describe('fixtures', () => {
-	for (const path of Object.keys(fixtures)) {
+	for (const path of Object.keys(fixturesTest)) {
 		const name = path.split('/').at(-1);
 		if (!name) {
 			continue;