+++ packages/skeleton-cli/build.config.ts
@@ -1,0 +1,7 @@
+import { defineBuildConfig } from 'unbuild';
+
+export default defineBuildConfig({
+	entries: ['./src/index.ts'],
+	declaration: true,
+	clean: true
+});