+++ packages/skeleton-cli/src/commands/migrate/migrations/skeleton-3/utility/constants.ts
@@ -1,0 +1,3 @@
+const FALLBACK_THEME = 'cerberus';
+
+export { FALLBACK_THEME };