+++ packages/skeleton-svelte/src/lib/components/Rating/Rating.svelte
@@ -17,7 +17,7 @@
 		controlGap = 'gap-2',
 		controlClasses,
 		// Label
-		labelBase = '',
+		labelBase = 'label-text',
 		labelClasses = '',
 		// Item
 		itemBase = '',