+++ packages/skeleton-svelte/src/lib/components/Tooltip/types.ts
@@ -6,9 +6,9 @@ export interface TooltipProps extends Omit<tooltip.Context, 'id' | 'open'> {
 	open?: boolean;
 
 	// Base ---
-	// Set base classes for the root element.
+	/** Set base classes for the root element. */
 	base?: string;
-	// Provide arbitrary classes for the root element.
+	/** Provide arbitrary classes for the root element. */
 	classes?: string;
 
 	// Trigger ---
@@ -36,8 +36,14 @@ export interface TooltipProps extends Omit<tooltip.Context, 'id' | 'open'> {
 	contentClasses?: string;
 
 	// Snippets ---
-	/** Provide the the trigger element, such as a button. */
+	/** Provide the template contents inside the trigger button. */
 	trigger?: Snippet;
 	/** Provide the template contents of the tooltip itself. */
 	content?: Snippet;
+
+	// Events ---
+	/** Handle the tooltip button hover event. */
+	onmouseover?: () => void;
+	/** Handle the tooltip button click event. */
+	onclick?: () => void;
 }