+++ sites/next.skeleton.dev/scripts/generate-schemas.ts
@@ -22,7 +22,16 @@ async function processFile(path: string): Promise<number> {
 	const outPath = join(OUTPUT_DIR, framework, `${toKebabCase(component)}.json`);
 	await mkdir(dirname(outPath), { recursive: true });
 	const interfaces = getInterfaces(path, {
-		matcher: /^[\w-]+Props$/
+		matcher: /^[\w-]+Props$/,
+		transformProperty(property) {
+			if (property.type.startsWith('Snippet')) {
+				return {
+					...property,
+					typeKind: 'primitive'
+				};
+			}
+			return property;
+		}
 	});
 	await writeFile(outPath, JSON.stringify(interfaces, null, 2));
 	return Object.keys(interfaces).length;